requirements.txt
asgiref==3.6.0
charset-normalizer==3.0.1
Django==4.2b1
django-filter==22.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
idna==3.4
Markdown==3.4.1
pymongo==3.12.3
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.2.4
tzdata==2022.7
urllib3==1.26.14
wincertstore==0.2
drf_spectacular
django-sslserver   
itsdangerous
python-dotenv
django-storages
Pillow
django-cors-headers 
django-seed
Faker
django-extensions==3.2.1
django-extensions==3.2.1
django-resized